Add ahothorization in order an owner of a restaurant to have access to delete and edit and an admin to delete it

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -3,13 +3,17 @@
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
+using Restaurants.Domain.Interfaces;
+using Restaurants.Domain.Constants;
 
 using MediatR;
+
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-                IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+                IRestaurantsRepository restaurantsRepository,
+                IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
         
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -19,6 +23,9 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             var restaurant = await restaurantsRepository.GetByIdAsync(request.Id);
             if (restaurant is null)
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await restaurantsRepository.Delete(restaurant);
